+++ packages/video_player/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.10.2
 
-* Adds optional VideoFormat used to signal what format the plugin should try.
+* **Android Only** Adds optional VideoFormat used to signal what format the plugin should try.
 
 ## 0.10.1+7